Move ServerInfo into Protocol, where its layer edges vanish
