+++ src/Protocol/InitializeResult.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Firehed\PhpLsp\Protocol;
 
-use Firehed\PhpLsp\ServerInfo;
 use JsonSerializable;
 
 /**
